Update Open-Gauge.ino
v0.10
Moved most stuff to functions to make it easier to tweak later (I hope)
Removed some values that were defined but as yet unused.
Changed the way that display elements are cleared. Not sure if it's more efficient or not though than clearing the entire display and re-drawing.
Stopped re-drawing UI boxes once they have been drawn.
Comments in lots of code now.
Altemeter to be removed later (I've also spelt it correctly this time)