Update Open-Gauge.ino
v0.17
Moved and fixed the way sensors are tested for at startup, also test for voltage read for voltmeter. If 0 we won't boot.
Added additional comments to help anyone following on from my code.
Added low voltage warning if < 12v - It's not good for batteries.
Added FAF mode due to a request.